MERGE: -c 61816 /branches/Platform-6.1 -> /trunk/Platform
Bug 13166 - [Mac10.7.2 -INTERNAL]LWISE : Password change failed for the domain user

10.7.2 is now doing its own password change handling based on the
value of the AuthenticationAuthority attribute, and the value we
have been returning, ";basic;", is not supported and causes an error.
We want it to just call into our DS plugin for password changes,
and the way to get it to do that is apparently to not return an
AuthenticationAuthority attribute at all.

CR: Kyle


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61817 9fc1c11e-11b2-4e14-94d8-6c5e05775969