Fix bug 13326 - "/opt/pbis" directory is not created after installing PBIS-OPEN
The install was successful, but uninstall was using the wrong fileset name to try to uninstall the packages. This prevents future installs from working.

The fix is to use - as a special subpackage name to indicate the fileset should match the package name, instead of appending something like .root at the end.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@62526 9fc1c11e-11b2-4e14-94d8-6c5e05775969