Fix bug 13346 - installation fails on FreeBSD 7.2 distro
On FreeBSD, libc has pthread stub functions. Those stub functions will check whether libthr is also linked and call the real function as necessary. The arguments are wrong for the pthread_mutexattr_settype function, and thus it fails.

The _pthread_mutexattr_settype function seems to work instead. So the fix is to additionally call it at runtime if necessary.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@62549 9fc1c11e-11b2-4e14-94d8-6c5e05775969