Fix bug 13839 - eventdbreaper generates warnings when parsing user monitor events from big endian machines
The eventdbreaper generated a warning because it could not decode the NDR encoded data field in the user monitor event.

The procedure NDR encoding happened to interoperate with Windows, but only from little endian systems. Big endian systems caused it to fail with error code 1730, RPC_S_UNSUPPORTED_TRANS_SYN, because Windows did not like the encoded NDR version after the NDR UUID.

The fix is to use Microsoft's format for the interface version and the transfer version for procedure message encoding.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@63899 9fc1c11e-11b2-4e14-94d8-6c5e05775969