Fix bug 13862 - user monitor segfault
The when looking up by SID, the LsaFindObjects command should return group objects, even for unenabled groups. However if it determines the group does not exist, instead of a more serious error, LsaFindObjects will return NULL at that location in the array.

Even though the user is a member of the SID (which means lsass knows about it), sometimes the group comes back as NULL.

The fix is to log an error and skip groups that stopped being resolvable. The memberships will be reported as deleted by usermonitor if they were enabled before.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@63926 9fc1c11e-11b2-4e14-94d8-6c5e05775969